+++ src/esrf_ontologies/technique/__init__.py
@@ -20,6 +20,10 @@ def get_techniques(*names: Tuple[str]) -> Set[Technique]:
     return set(_iter_from_names(*names))
 
 
+def get_ontology_version(metadata: TechniqueMetadata) -> str:
+    return next(iter(metadata.techniques)).ontology_version
+
+
 @lru_cache(maxsize=1)
 def get_all_techniques() -> List[Technique]:
     """Returns a list of techniques."""