+++ src/esrf_ontologies/tests/test_single_technique_metadata.py
@@ -1,13 +1,16 @@
 import pytest
 
+from ..technique import get_ontology_version
 from ..technique import get_technique_metadata
 
 
 def test_get_dataset_metadata():
     metadata = get_technique_metadata("XAS")
+    ontology_version = get_ontology_version(metadata)
     dataset_metadata = {
         "definition": "XAS",
         "technique_pid": "https://w3id.org/PaN/ESRFET#XAS",
+        "technique_pid_esrfet_version": ontology_version,
     }
 
     assert metadata.get_dataset_metadata() == dataset_metadata
@@ -15,9 +18,11 @@ def test_get_dataset_metadata():
 
 def test_fill_dataset_metadata():
     metadata = get_technique_metadata("XAS")
+    ontology_version = get_ontology_version(metadata)
     dataset_metadata = {
         "definition": "XAS",
         "technique_pid": "https://w3id.org/PaN/ESRFET#XAS",
+        "technique_pid_esrfet_version": ontology_version,
     }
 
     dataset = {}